% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatsTrafego.R
\name{computeStatsTrafego}
\alias{computeStatsTrafego}
\title{Compute stats for ride traffic}
\usage{
computeStatsTrafego(rides, nome_variavel)
}
\arguments{
\item{rides}{a data.frame/data.table object with ride information from Strava.}

\item{nome_variavel}{a character object with the variable name to be analyzed.}
}
\value{
a list object with statistics for streets categorized as low, medium 
and high traffic.
}
\description{
This function will compute a number of statistics about the ride traffic 
for the specified nome_variable
}
