% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_strava.R
\name{processing_strava}
\alias{processing_strava}
\title{Processing data from Strava dataset}
\usage{
processing_strava(metadata, shape, dateVar)
}
\arguments{
\item{metadata}{data.frame containing metadata about the rides.}

\item{shape}{data.frame containing the geographic features (shapefile) of the rides.}

\item{dataVar}{character name of the column representing the "period" of the metadata file. 
For files from all_edges_monthly_*, this accounts for "month".}
}
\value{
a data.frame with the metadata merged with its geographic features.
}
\description{
This function will process the monthly dataset files from Strava Metro. Specifically, 
it will process the metadata file and merge with the related shapefile.
}
\details{
The information provided by Strava Metro is a set of directory containing data 
about all rides in hourly, daily and monthly period. This function is aimed to
be used with the monthly data, that is, files contained in the 
all_edges_monthly_* directory.
}
