% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialMerge_stravaMalhaPermanente.R
\name{spatialMerge_stravaMalhaPermanente}
\alias{spatialMerge_stravaMalhaPermanente}
\title{Spatial merge between Strava and Cycling Network}
\usage{
spatialMerge_stravaMalhaPermanente(strava, malhaPermanente)
}
\arguments{
\item{strava}{data.frame/data.table object with ride information from Strava. 
This is the result from the processing_strava() function.}

\item{malhaPermanente}{data.frame with the geographical features from the 
cycling network. Basically, it is the object resulted from reading the .shp file
using the sf package.}
}
\value{
a data.table object with ride information merged with 
geographical features from the cycling network.
}
\description{
This function will perform a spatial merge between dataset from Strava Metro
and the cycling network.
}
\details{
The spatial merge is done using a buffering of the cycling network (15 meters),
and then joining it with ride information from Strava by using the 'within distance (5 meters)'
criteria. This process is able to identify the rides that are covered by the
cycling network.
}
